+++ packages/_private/e2e-demo-test/src/test.selector.ts
@@ -36,6 +36,7 @@ const AllComponentSelectors = {
   input417NumericModern: '#tab-content-Input syn-input[data-testid="input-417-numeric-modern"]',
   input417NumericNative: '#tab-content-Input syn-input[data-testid="input-417-numeric-native"]',
   input872SpinButtons: '#tab-content-Input syn-input[data-testid="input-872-spin-buttons"]',
+  input1023Autocorrect: '#tab-content-Input syn-input[data-testid="input-1023-autocorrect"]',
   inputContent: '#tab-content-Input',
   inputLink: '#tab-Input',
 
@@ -65,6 +66,7 @@ const AllComponentSelectors = {
   select847Multiple: '#tab-content-Select syn-select[data-testid="select-847-multiple"]',
   select885ValueZeroNumber: '#tab-content-Select syn-select[data-testid="select-885-value-zero-number"]',
   select885ValueZeroString: '#tab-content-Select syn-select[data-testid="select-885-value-zero-string"]',
+  select1036Delimiter: '#tab-content-Select syn-select[data-testid="select-1036-subsequently-changed-delimiter"]',
   selectContent: '#tab-content-Select',
   selectForm: '#tab-content-Select syn-select[data-testid="select-form-813"]',
   selectFormOptions: '#tab-content-Select syn-select[data-testid="select-form-813"] syn-option',