+++ packages/angular/components/combobox/combobox.component.ts
@@ -246,6 +246,8 @@ and `hide()` methods and this attribute will reflect the combobox's open state.
 * Enable this option to prevent the listbox from being clipped,
 when the component is placed inside a container with `overflow: auto|scroll`.
 Hoisting uses a fixed positioning strategy that works in many, but not all, scenarios.
+
+@deprecated This property is deprecated and will be removed in the next major version.
  */
   @Input()
   set hoist(v: '' | SynCombobox['hoist']) {