+++ packages/docs/.storybook/preview.ts
@@ -23,6 +23,7 @@ import { LIGHT_THEME, DARK_THEME, SICK_2025_DARK, SICK_2025_LIGHT, Chromatic_Mod
 import { generateFigmaPluginObject } from '../src/helpers/figma.js';
 import docsCodepenEnhancer from '../src/docs-codepen-enhancer/index.js';
 import { themeSwitchIcons } from '../src/decorators/ThemeSwitchIcons.js';
+import { themeNotSupportedHint } from '../src/decorators/ThemeNotSupportedHint.js';
 
 // Filter out all private members and readonly properties from the manifest
 const filteredManifest = (manifest: Package): Package => ({
@@ -70,7 +71,7 @@ const themeByClassName = withThemeByClassName<WebComponentsRenderer>({
 });
 
 const preview: Preview = {
-  decorators: [stopAnimation, themeByClassName, themeSwitchIcons],
+  decorators: [stopAnimation, themeByClassName, themeSwitchIcons, themeNotSupportedHint],
   initialGlobals: {
     background: {
       value: 'neutral-50',
@@ -104,7 +105,7 @@ const preview: Preview = {
       stories: { inline: false },
       toc: {
         headingSelector: 'h2, h3',
-        ignoreSelector: '.toc-ignore h2, .toc-ignore h3',
+        ignoreSelector: '.toc-ignore h2, .toc-ignore h3, .docs-story h2, .docs-story h3',
       },
       source: {
         format: 'html',