+++ packages/docs/stories/components/badge.stories.ts
@@ -41,7 +41,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Feedback'],
+  tags: ['Feedback', 'SICK2018'],
   title: 'Components/syn-badge',
 };
 export default meta;