+++ packages/docs/stories/components/menu-item.stories.ts
@@ -38,7 +38,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Navigation'],
+  tags: ['Navigation', 'SICK2018'],
   title: 'Components/syn-menu-item',
 };
 export default meta;