+++ packages/docs/stories/components/range.stories.ts
@@ -39,7 +39,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Form'],
+  tags: ['Form', 'SICK2018'],
   title: 'Components/syn-range',
 };
 export default meta;