+++ packages/docs/stories/components/select.stories.ts
@@ -40,7 +40,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Form'],
+  tags: ['Form', 'SICK2018'],
   title: 'Components/syn-select',
 };
 export default meta;