+++ packages/docs/stories/components/side-nav.stories.tsx
@@ -33,7 +33,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Application Shell'],
+  tags: ['Application Shell', 'SICK2018'],
   title: 'Components/syn-side-nav',
 };