+++ packages/docs/stories/components/tab.stories.ts
@@ -31,7 +31,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Navigation'],
+  tags: ['Navigation', 'SICK2018'],
   title: 'Components/syn-tab',
 };
 export default meta;