+++ packages/docs/stories/components/tag.stories.ts
@@ -13,6 +13,7 @@ import {
   storybookTemplate,
 } from '../../src/helpers/component.js';
 import { generateFigmaPluginObject } from '../../src/helpers/figma.js';
+import { Chromatic_Modes_All } from '../../.storybook/modes.js';
 
 const { args: defaultArgs, argTypes } = storybookDefaults('syn-tag');
 const { generateTemplate } = storybookTemplate('syn-tag');
@@ -25,6 +26,9 @@ const meta: Meta = {
   argTypes,
   component: 'syn-tag',
   parameters: {
+    chromatic: {
+      modes: Chromatic_Modes_All,
+    },
     design: generateFigmaPluginObject('2898-58777'),
     docs: {
       description: {
@@ -32,7 +36,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Feedback'],
+  tags: ['Feedback', 'SICK2018', 'SICK2025'],
   title: 'Components/syn-tag',
 };
 export default meta;