+++ packages/docs/stories/styles/table.stories.ts
@@ -120,7 +120,7 @@ const meta: Meta = {
       },
     },
   },
-  tags: ['Structure', 'Styles', 'SICK2025'],
+  tags: ['Structure', 'Styles', 'SICK2018', 'SICK2025'],
   title: 'Styles/syn-table',
 };
 export default meta;